COPYING: fix incorrect FSF address
This replaces COPYING with the current text from:
https://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt

The address of the Free Software Foundation is updated.

This also introduces some trivial whitespace changes which would
not ordinarily be wanted in a patch, but it is nicer for downstream
distributors if the COPYING file exactly matches the "standard" LGPL
license text that is used in other packages.